+++ frontend/src/components/@shared/LinkText/type.ts
