+++ frontend/src/components/@shared/ScrollToTop/type.ts
