+++ frontend/src/components/index.ts
@@ -10,7 +10,9 @@ export * from 'components/@shared/Dropdown';
 export * from 'components/@shared/Input';
 export * from 'components/@shared/LoadingButton';
 export * from 'components/@shared/LoadingSpinner';
-export * from 'components/LinkText';
+export * from 'components/@shared/Popover';
+export * from 'components/@shared/LinkText';
+export * from 'components/@shared/ScrollToTop';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
 export * from 'components/Header';
@@ -46,8 +48,6 @@ export * from 'components/ErrorFallbackMain';
 export * from 'components/ErrorFallbackLoginButton';
 export * from 'components/ErrorFallbackSubscriptionButton';
 export * from 'components/ErrorFallbackNotificationMessage';
-export * from 'components/ScrollToTop';
-export * from 'components/Popover';
 export * from 'components/MembersPopover';
 export * from 'components/CheckSuccessCycle';
 export * from 'components/Tabs';
